+++ src/main/java/com/nhs2304/demosortalgo/hello/HelloController.java
@@ -1,4 +1,4 @@
-package com.nhs2304.demosortalgo;
+package com.nhs2304.demosortalgo.hello;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
